refactor: Introduce aliases
