+++ configurations/css/plugins/_index.ts
@@ -1,7 +1,7 @@
-export { default as pluginDeclarationBlockNoIgnoredProperties } from './declaration-block-no-ignored-properties.js';
-export { default as pluginHighPerformanceAnimation } from './high-performance-animation.js';
-export { default as pluginNoNestedMedia } from './no-nested-media.js';
-export { default as pluginOrder } from './order.js';
-export { default as pluginSelectorNoEmpty } from './selector-no-empty.js';
-export { default as pluginSelectorTagNoWithoutClass } from './selector-tag-no-without-class.js';
-export { default as pluginStylistic } from './stylistic.js';
+export { default as pluginDeclarationBlockNoIgnoredProperties } from './declaration-block-no-ignored-properties';
+export { default as pluginHighPerformanceAnimation } from './high-performance-animation';
+export { default as pluginNoNestedMedia } from './no-nested-media';
+export { default as pluginOrder } from './order';
+export { default as pluginSelectorNoEmpty } from './selector-no-empty';
+export { default as pluginSelectorTagNoWithoutClass } from './selector-tag-no-without-class';
+export { default as pluginStylistic } from './stylistic';
