+++ configurations/css/rules/_index.ts
@@ -1,2 +1,2 @@
-export { default as stylelintAvoidErrors } from './avoid-errors.js';
-export { default as stylelintEnforceConventions } from './enforce-conventions.js';
+export { default as stylelintAvoidErrors } from './avoid-errors';
+export { default as stylelintEnforceConventions } from './enforce-conventions';
