+++ configurations/index.ts
@@ -1,2 +1,2 @@
-export { default as css } from './css/index.js';
-export { default as scss } from './scss/index.js';
+export { default as css } from './css/index';
+export { default as scss } from './scss/index';
