+++ tsconfig.json
@@ -5,10 +5,14 @@
 		"esModuleInterop": true,
 		"exactOptionalPropertyTypes": false,
 		"lib": ["ES2022"],
-		"module": "NodeNext",
-		"moduleResolution": "NodeNext",
+		"module": "ESNext",
+		"moduleResolution": "bundler",
 		"noEmit": true,
 		"noUncheckedIndexedAccess": false,
+		"paths": {
+			"~shared/*": ["./shared/*"],
+			"~utils/*": ["./utils/*"],
+		},
 		"resolveJsonModule": true,
 		"skipLibCheck": true,
 		"strict": true,
